Implement UserForm component for Profile2 page
